Routing

+++ routes/reviewRoute.js
@@ -3,19 +3,17 @@ const router = express.Router()
 const catchAsync = require('../utils/catchAsync')
 const Review = require('../models/Review')
 const Recipe = require('../models/Recipe')
-
-router.post('/recipe/:id/reviews',catchAsync(async(req,res)=>{ 
-    const {id} = req.params 
-     const recipe = await Recipe.findById(id)
-     const review = new Review(req.body) 
-     recipe.reviews.push(review)
-     await   review.save()
- 
-     await  recipe.save()
-     res.redirect(`/recipe/${recipe._id}`) 
- }))
- router.delete('/recipe/:id/reviews/:reviewId',catchAsync(async(req,res)=>{ 
+const isLoggedIn = require('../middleware')
+const {postReview} = require('../controllers/Reviews')
+router.post('/recipe/:id/reviews',isLoggedIn,postReview)
+ router.delete('/recipe/:id/reviews/:reviewId',isLoggedIn,catchAsync(async(req,res)=>{ 
      const {id,reviewId} = req.params
+     const review = await Review.findById(reviewId) 
+     if(!review.author.equals(req.user._id))
+     { 
+        req.flash('error',"You cannot do that")
+        return res.redirect(`/recipe/${id}`)
+     }
       await Recipe.findByIdAndUpdate(id,{$pull:{reviews:reviewId}})
      await Review.findByIdAndDelete(reviewId) 
      res.redirect(`/recipe/${id}`)
